v4.0 A.3: split mcp/tools.py per domain
Extracted the 1278-line MCP tools module into focused per-domain
files. Shared backend-context/REST plumbing centralized in
_runtime.py. Public API (execute_tool, TOOLS registry) preserved
for stdio + HTTP/SSE transports.

- mnemos/mcp/tools/__init__.py: registry assembly + dispatcher
- mnemos/mcp/tools/_runtime.py: REST plumbing + auth helpers
- mnemos/mcp/tools/memory.py: search/get/list/create/update/delete/bulk
- mnemos/mcp/tools/kg.py: KG triple operations
- mnemos/mcp/tools/dag.py: log/branch/diff/checkout
- mnemos/mcp/tools/models.py: recommend_model

No wire-shape changes. 925 unit tests pass.

Authored-By: Codex
