+++ tests/test_mcp_stdio_wire.py
@@ -16,13 +16,17 @@ from pathlib import Path
 from typing import Iterable
 
 REPO_ROOT = Path(__file__).parent.parent
-MCP_TOOLS = REPO_ROOT / "mnemos" / "mcp" / "tools.py"
+MCP_TOOLS = REPO_ROOT / "mnemos" / "mcp" / "tools"
 
 
 def _extract_mcp_paths() -> list[str]:
     """Pull every literal/f-string path passed to _rest_* in
-    mnemos/mcp/tools.py. Returns the static prefix up to the first f-string hole."""
-    src = MCP_TOOLS.read_text(encoding="utf-8")
+    mnemos/mcp/tools/. Returns the static prefix up to the first f-string hole."""
+    src = "\n".join(
+        path.read_text(encoding="utf-8")
+        for path in sorted(MCP_TOOLS.glob("*.py"))
+        if path.name != "__init__.py"
+    )
     # Match _rest_get("..."), _rest_post("..."), _rest_delete("..."), and f"..." variants.
     # We capture the leading literal segment; f-strings with `{arg}` holes
     # keep only the static prefix so we compare against registered path
@@ -42,7 +46,7 @@ def _extract_mcp_paths() -> list[str]:
 def _registered_prefixes() -> list[str]:
     """Enumerate the static prefixes of every route registered in the
     FastAPI app. Returns prefixes suitable for startswith() matching
-    against mnemos/mcp/tools.py's literal path prefixes."""
+    against mnemos/mcp/tools/ literal path prefixes."""
     # Import lazily — the app import chain pulls in asyncpg / pgvector
     # stubs, which conftest.py sets up.
     from mnemos.api.main import app  # noqa: E402
@@ -81,8 +85,8 @@ class TestMCPWireContract:
     def test_every_mcp_path_is_a_real_route(self):
         mcp_paths = _extract_mcp_paths()
         registered = _registered_prefixes()
-        assert mcp_paths, "mnemos/mcp/tools.py exposes no paths — something is wrong with extraction"
-        assert registered, "FastAPI app exposes no routes — import failed"
+        assert mcp_paths, "mnemos/mcp/tools exposes no paths - something is wrong with extraction"
+        assert registered, "FastAPI app exposes no routes - import failed"
 
         missing: list[str] = []
         for path in mcp_paths:
@@ -90,22 +94,22 @@ class TestMCPWireContract:
                 missing.append(path)
 
         assert not missing, (
-            f"mnemos/mcp/tools.py calls these paths that the REST app does not serve: {missing}. "
-            f"This is the #M31-01 regression — MCP stdio server returns 404 for callers. "
+            f"mnemos/mcp/tools calls these paths that the REST app does not serve: {missing}. "
+            f"This is the #M31-01 regression - MCP stdio server returns 404 for callers. "
             f"Registered prefixes (sample): {sorted(registered)[:20]}"
         )
 
     def test_memory_paths_are_v1_prefixed(self):
-        """Explicit regression: every memory-related path in mnemos/mcp/tools.py
+        """Explicit regression: every memory-related path in mnemos/mcp/tools
         must carry the /v1 prefix. Reverting the prefix would reintroduce
         the v3.0.0 bug."""
         mcp_paths = _extract_mcp_paths()
         memory_paths = [p for p in mcp_paths if "/memor" in p]
-        assert memory_paths, "mnemos/mcp/tools.py has no memory paths — extraction regex is broken"
+        assert memory_paths, "mnemos/mcp/tools has no memory paths - extraction regex is broken"
 
         unprefixed = [p for p in memory_paths if not p.startswith("/v1/")]
         assert not unprefixed, (
-            f"These memory paths in mnemos/mcp/tools.py are missing the /v1 prefix: {unprefixed}. "
+            f"These memory paths in mnemos/mcp/tools are missing the /v1 prefix: {unprefixed}. "
             f"The REST router registers memories under /v1 (api/handlers/memories.py:34); "
             f"any path without the prefix will 404."
         )
@@ -113,18 +117,18 @@ class TestMCPWireContract:
     def test_kg_paths_are_v1_prefixed(self):
         """KG routes were moved from /kg to /v1/kg as part of the v3.3
         API cleanup pass. Any
-        unversioned /kg/* path in mnemos/mcp/tools.py would 404 against the
+        unversioned /kg/* path in mnemos/mcp/tools would 404 against the
         current API."""
         mcp_paths = _extract_mcp_paths()
         kg_paths = [p for p in mcp_paths if "/kg/" in p or p.endswith("/kg")]
-        assert kg_paths, "mnemos/mcp/tools.py has no KG paths — extraction regex is broken"
+        assert kg_paths, "mnemos/mcp/tools has no KG paths - extraction regex is broken"
 
         unprefixed = [
             p for p in kg_paths
             if (p.startswith("/kg/") or p == "/kg") and not p.startswith("/v1/")
         ]
         assert not unprefixed, (
-            f"These KG paths in mnemos/mcp/tools.py are missing the /v1 prefix: "
+            f"These KG paths in mnemos/mcp/tools are missing the /v1 prefix: "
             f"{unprefixed}. The REST router registers KG under /v1/kg "
             f"(api/handlers/kg.py); any path without the prefix will 404."
         )
